DayOne task 1
The task is to take a data object and convert it into two arrays based
on the column.

Then sort the array and calculate the distance (int) betweent the two
numbers.

Since the integer is representing a distance, you need to subtract the
highest int from the lowest int to ensure you don't get minus numbers.